requirements.txt
pytest
pydantic
requests
allure-pytest
allure-python-commons